Decoupled service into service and utils.
